1, update todo list 2, change serverid from uint16 to uint8 3, add gateservercontext on master server 4, update loginservercontext and gateservercontext transfer
vet/app: me

+++ trunk/MasterServer/master_logic_command.h
@@ -54,7 +54,7 @@ struct LogicCommandGateAllocAck : public LogicCommandT<COMMAND_GATEALLOCACK>
 
 	uint32 m_iLoginSessionId;
 	uint32 m_iGateSessionId;
-	uint16 m_iServerId;
+	uint8 m_iServerId;
 	TCHAR m_strAccountName[ACCOUNTNAME_MAX + 1];
 };
